Move service selector into component
