+++ apps/site/assets/ts/schedule/components/schedule-finder/ScheduleModalContent.tsx
@@ -4,16 +4,7 @@ import UpcomingDepartures from "./UpcomingDepartures";
 import { Route, RouteType, ServiceWithServiceDate } from "../../../__v3api";
 import { SimpleStop, StopPrediction } from "../__schedule";
 import isSilverLine from "../../../helpers/silver-line";
-import {
-  serviceDays,
-  groupServiceByDate,
-  ServiceByOptGroup,
-  ServiceOptGroup,
-  ServicesKeyedByGroup,
-  groupByType,
-  getTodaysSchedule
-} from "../../../helpers/service";
-import SelectContainer from "./SelectContainer";
+import { ServiceSelector } from "./ServiceSelector";
 
 const stopInfo = (
   selectedOrigin: string,
@@ -128,18 +119,6 @@ const ScheduleModalContent = ({
     return null;
   }
   const destination = directionDestinations[selectedDirection];
-  const servicesByOptGroup: ServicesKeyedByGroup = services
-    .map((service: ServiceWithServiceDate) => groupServiceByDate(service))
-    .reduce(groupByType, { current: [], holiday: [], other: [] });
-
-  const todayService = getTodaysSchedule(servicesByOptGroup);
-  const defaultServiceId = todayService ? todayService.service.id : "";
-  const optGroupNames: ServiceOptGroup[] = ["current", "holiday", "other"];
-  const optGroupTitles: { [key in ServiceOptGroup]: string } = {
-    current: "Current Schedules",
-    holiday: "Holiday Schedules",
-    other: "Other Schedules"
-  };
   return (
     <>
       <div className="schedule-finder__modal-header">
@@ -156,32 +135,7 @@ const ScheduleModalContent = ({
       </div>
       <div>from {stopNameLink(selectedOrigin, stops)}</div>
       <UpcomingDepartures state={state} />
-
-      <div className="schedule-finder__service-selector">
-        <SelectContainer id="service_selector_container" error={false}>
-          <select
-            id="service_selector"
-            className="schedule-finder__select"
-            defaultValue={defaultServiceId}
-          >
-            {optGroupNames.map((group: ServiceOptGroup) => (
-              <optgroup key={group} label={optGroupTitles[group]}>
-                {servicesByOptGroup[group].map((service: ServiceByOptGroup) => {
-                  const daysOfWeek = serviceDays(service.service);
-                  return (
-                    <option value={service.service.id} key={service.service.id}>
-                      {service.service.description}
-                      {daysOfWeek && ` ${daysOfWeek}`}
-                      {group !== "holiday" ? ", " : " "}
-                      {service.servicePeriod}
-                    </option>
-                  );
-                })}
-              </optgroup>
-            ))}
-          </select>
-        </SelectContainer>
-      </div>
+      <ServiceSelector services={services} />
     </>
   );
 };
